fix: split debug subcommands into separate files per codegen
Each debug subcommand must be in its own file matching the codegen
output structure, not combined into parent files.
